PHPLIB-252: Monitor find commands to test ReadableStream::seek()
The previous tests neglected to assert the bytes that were actually read after seeking. Additionally, the test for seeking to a subsequent chunk asserted that a mocked, private method was never called, which always passed as PHPUnit cannot track private methods.

This revises the tests to use command monitoring to assert exactly how many queries should be should by a seek (if at all).